requirements.txt
asgiref==3.3.4
Django==3.2.3
django-environ==0.4.5
django-extensions==3.1.3
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
